Refactor post validation
