+++ client/modules/post/validation/edit.ts
@@ -1,0 +1,6 @@
+import { createPostValidationSchema } from "@/modules/post/validation/create"
+import { z } from "zod"
+
+const editPostValidationSchema = z.object({}).merge(createPostValidationSchema)
+
+export { editPostValidationSchema }
